refactor(air): move SymbolicAirBuilder from uni-stark to air crate
Move the core `SymbolicAirBuilder` struct and constraint evaluation
functions to `p3_air::symbolic`. This enables other crates to use
symbolic constraint analysis without depending on uni-stark.

STARK-specific quotient degree calculations (`get_log_num_quotient_chunks`,
`get_log_quotient_degree_extension`) remain in uni-stark.
