+++ air/Cargo.toml
@@ -13,6 +13,7 @@ categories.workspace = true
 p3-field.workspace = true
 p3-matrix.workspace = true
 serde = { workspace = true, features = ["derive", "alloc"] }
+tracing.workspace = true
 
 [dev-dependencies]
 p3-baby-bear = { path = "../baby-bear" }
