+++ air/src/symbolic/mod.rs
@@ -1,7 +1,9 @@
 //! Symbolic expression types for AIR constraint representation.
 
+mod builder;
 mod expression;
 mod variable;
 
+pub use builder::*;
 pub use expression::SymbolicExpression;
 pub use variable::{Entry, SymbolicVariable};
