+++ batch-stark/src/prover.rs
@@ -4,6 +4,7 @@ use alloc::vec::Vec;
 use p3_air::Air;
 #[cfg(debug_assertions)]
 use p3_air::DebugConstraintBuilder;
+use p3_air::symbolic::{SymbolicAirBuilder, SymbolicExpression};
 use p3_challenger::{CanObserve, FieldChallenger};
 use p3_commit::{Pcs, PolynomialSpace};
 use p3_field::{BasedVectorSpace, PackedFieldExtension, PackedValue, PrimeCharacteristicRing};
@@ -13,10 +14,7 @@ use p3_lookup::lookup_traits::{Kind, Lookup, LookupData, LookupGadget, lookup_da
 use p3_matrix::Matrix;
 use p3_matrix::dense::RowMajorMatrix;
 use p3_maybe_rayon::prelude::*;
-use p3_uni_stark::{
-    OpenedValues, PackedChallenge, PackedVal, ProverConstraintFolder, SymbolicAirBuilder,
-    SymbolicExpression,
-};
+use p3_uni_stark::{OpenedValues, PackedChallenge, PackedVal, ProverConstraintFolder};
 use p3_util::log2_strict_usize;
 use tracing::{debug_span, info_span, instrument};
