+++ batch-stark/src/verifier.rs
@@ -5,6 +5,7 @@ use core::fmt::Debug;
 
 use hashbrown::HashMap;
 use p3_air::Air;
+use p3_air::symbolic::{SymbolicAirBuilder, SymbolicExpression};
 use p3_challenger::{CanObserve, FieldChallenger};
 use p3_commit::{Pcs, PolynomialSpace};
 use p3_field::{BasedVectorSpace, PrimeCharacteristicRing};
@@ -15,10 +16,7 @@ use p3_lookup::lookup_traits::{
 };
 use p3_matrix::dense::RowMajorMatrixView;
 use p3_matrix::stack::VerticalPair;
-use p3_uni_stark::{
-    SymbolicAirBuilder, SymbolicExpression, VerificationError, VerifierConstraintFolder,
-    recompose_quotient_from_chunks,
-};
+use p3_uni_stark::{VerificationError, VerifierConstraintFolder, recompose_quotient_from_chunks};
 use p3_util::zip_eq::zip_eq;
 use tracing::{info_span, instrument};
