+++ examples/src/airs.rs
@@ -1,3 +1,4 @@
+use p3_air::symbolic::SymbolicAirBuilder;
 use p3_air::{Air, AirBuilder, BaseAir};
 use p3_blake3_air::Blake3Air;
 use p3_challenger::FieldChallenger;
@@ -8,8 +9,7 @@ use p3_matrix::dense::RowMajorMatrix;
 use p3_poseidon2::GenericPoseidon2LinearLayers;
 use p3_poseidon2_air::{Poseidon2Air, VectorizedPoseidon2Air};
 use p3_uni_stark::{
-    DebugConstraintBuilder, ProverConstraintFolder, StarkGenericConfig, SymbolicAirBuilder,
-    VerifierConstraintFolder,
+    DebugConstraintBuilder, ProverConstraintFolder, StarkGenericConfig, VerifierConstraintFolder,
 };
 use rand::distr::StandardUniform;
 use rand::prelude::Distribution;
