+++ uni-stark/src/preprocessed.rs
@@ -1,9 +1,10 @@
 use p3_air::Air;
+use p3_air::symbolic::SymbolicAirBuilder;
 use p3_commit::Pcs;
 use p3_matrix::Matrix;
 use tracing::debug_span;
 
-use crate::{ProverConstraintFolder, StarkGenericConfig, SymbolicAirBuilder, Val};
+use crate::{ProverConstraintFolder, StarkGenericConfig, Val};
 
 /// Prover-side reusable data for preprocessed columns.
 ///
